+++ android/SSPSDK.kt
@@ -1030,4 +1030,4 @@ sealed class SDKError : Exception() {
 }
 
 // SDK version
-const val SDK_VERSION = "0.5.25"
+const val SDK_VERSION = "0.5.26"